+++ volume/export.go
@@ -1,0 +1,230 @@
+/*
+Copyright 2016 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package volume
+
+import (
+	"fmt"
+	"io/ioutil"
+	"os/exec"
+	"regexp"
+	"strconv"
+	"strings"
+	"sync"
+
+	"github.com/golang/glog"
+	"github.com/guelfey/go.dbus"
+	"k8s.io/client-go/pkg/api/v1"
+)
+
+type exporter interface {
+	AddExportBlock(string) (string, uint16, error)
+	RemoveExportBlock(string, uint16) error
+	Export(string) error
+	Unexport(*v1.PersistentVolume) error
+}
+
+type exportBlockCreator interface {
+	CreateExportBlock(string, string) string
+}
+
+type genericExporter struct {
+	ebc    exportBlockCreator
+	config string
+
+	// Map to track used exportIds. Each ganesha export needs a unique fsid and
+	// Export_Id, each kernel a unique fsid. Assign each export an exportId and
+	// use it as both fsid and Export_Id.
+	exportIds map[uint16]bool
+
+	mapMutex  *sync.Mutex
+	fileMutex *sync.Mutex
+}
+
+func newGenericExporter(ebc exportBlockCreator, config string, re *regexp.Regexp) *genericExporter {
+	exportIds, err := getExistingIds(config, re)
+	if err != nil {
+		glog.Errorf("error while populating exportIds map, there may be errors exporting later if exportIds are reused: %v", err)
+	}
+	return &genericExporter{
+		ebc:       ebc,
+		config:    config,
+		exportIds: exportIds,
+		mapMutex:  &sync.Mutex{},
+		fileMutex: &sync.Mutex{},
+	}
+}
+
+// getExistingIds populates an exportIds map with pre-existing exportIds
+// found in the given config file. Takes as argument the regex it should use to
+// find each exportId in the file i.e. Export_Id or fsid.
+func getExistingIds(config string, re *regexp.Regexp) (map[uint16]bool, error) {
+	exportIds := map[uint16]bool{}
+
+	digitsRe := "([0-9]+)"
+	if !strings.Contains(re.String(), digitsRe) {
+		return exportIds, fmt.Errorf("regexp %s doesn't contain digits submatch %s", re.String(), digitsRe)
+	}
+
+	read, err := ioutil.ReadFile(config)
+	if err != nil {
+		return exportIds, err
+	}
+
+	allMatches := re.FindAllSubmatch(read, -1)
+	for _, match := range allMatches {
+		digits := match[1]
+		if id, err := strconv.ParseUint(string(digits), 10, 16); err == nil {
+			exportIds[uint16(id)] = true
+		}
+	}
+
+	return exportIds, nil
+}
+
+func (e *genericExporter) CreateExportBlock(_, _ string) string {
+	return ""
+}
+
+func (e *genericExporter) AddExportBlock(path string) (string, uint16, error) {
+	exportId := generateId(e.mapMutex, e.exportIds)
+	exportIdStr := strconv.FormatUint(uint64(exportId), 10)
+
+	block := e.ebc.CreateExportBlock(exportIdStr, path)
+
+	// Add the export block to the config file
+	if err := addToFile(e.fileMutex, e.config, block); err != nil {
+		deleteId(e.mapMutex, e.exportIds, exportId)
+		return "", 0, fmt.Errorf("error adding export block %s to config %s: %v", block, e.config, err)
+	}
+	return block, exportId, nil
+}
+
+func (e *genericExporter) RemoveExportBlock(block string, exportId uint16) error {
+	deleteId(e.mapMutex, e.exportIds, exportId)
+	return removeFromFile(e.fileMutex, e.config, block)
+}
+
+type ganeshaExporter struct {
+	genericExporter
+}
+
+var _ exporter = &ganeshaExporter{}
+
+func newGaneshaExporter(ganeshaConfig string) exporter {
+	return &ganeshaExporter{
+		genericExporter: *newGenericExporter(&ganeshaExportBlockCreator{}, ganeshaConfig, regexp.MustCompile("Export_Id = ([0-9]+);")),
+	}
+}
+
+// Export exports the given directory using NFS Ganesha, assuming it is running
+// and can be connected to using D-Bus.
+func (e *ganeshaExporter) Export(path string) error {
+	// Call AddExport using dbus
+	conn, err := dbus.SystemBus()
+	if err != nil {
+		return fmt.Errorf("error getting dbus session bus: %v", err)
+	}
+	obj := conn.Object("org.ganesha.nfsd", "/org/ganesha/nfsd/ExportMgr")
+	call := obj.Call("org.ganesha.nfsd.exportmgr.AddExport", 0, e.config, fmt.Sprintf("export(path = %s)", path))
+	if call.Err != nil {
+		return fmt.Errorf("error calling org.ganesha.nfsd.exportmgr.AddExport: %v", call.Err)
+	}
+
+	return nil
+}
+
+func (e *ganeshaExporter) Unexport(volume *v1.PersistentVolume) error {
+	ann, ok := volume.Annotations[annExportId]
+	if !ok {
+		return fmt.Errorf("PV doesn't have an annotation %s, can't remove the export from the server", annExportId)
+	}
+	exportId, _ := strconv.ParseUint(ann, 10, 16)
+
+	// Call RemoveExport using dbus
+	conn, err := dbus.SystemBus()
+	if err != nil {
+		return fmt.Errorf("error getting dbus session bus: %v", err)
+	}
+	obj := conn.Object("org.ganesha.nfsd", "/org/ganesha/nfsd/ExportMgr")
+	call := obj.Call("org.ganesha.nfsd.exportmgr.RemoveExport", 0, uint16(exportId))
+	if call.Err != nil {
+		return fmt.Errorf("error calling org.ganesha.nfsd.exportmgr.RemoveExport: %v", call.Err)
+	}
+
+	return nil
+}
+
+type ganeshaExportBlockCreator struct{}
+
+var _ exportBlockCreator = &ganeshaExportBlockCreator{}
+
+// CreateBlock creates the text block to add to the ganesha config file.
+func (e *ganeshaExportBlockCreator) CreateExportBlock(exportId, path string) string {
+	return "\nEXPORT\n{\n" +
+		"\tExport_Id = " + exportId + ";\n" +
+		"\tPath = " + path + ";\n" +
+		"\tPseudo = " + path + ";\n" +
+		"\tAccess_Type = RW;\n" +
+		"\tSquash = root_id_squash;\n" +
+		"\tSecType = sys;\n" +
+		"\tFilesystem_id = " + exportId + "." + exportId + ";\n" +
+		"\tFSAL {\n\t\tName = VFS;\n\t}\n}\n"
+}
+
+type kernelExporter struct {
+	genericExporter
+}
+
+var _ exporter = &kernelExporter{}
+
+func newKernelExporter() exporter {
+	return &kernelExporter{
+		genericExporter: *newGenericExporter(&kernelExportBlockCreator{}, "/etc/exports", regexp.MustCompile("fsid=([0-9]+)")),
+	}
+}
+
+// Export exports all directories listed in /etc/exports
+func (e *kernelExporter) Export(_ string) error {
+	// Execute exportfs
+	cmd := exec.Command("exportfs", "-r")
+	out, err := cmd.CombinedOutput()
+	if err != nil {
+		return fmt.Errorf("exportfs -r failed with error: %v, output: %s", err, out)
+	}
+
+	return nil
+}
+
+func (e *kernelExporter) Unexport(volume *v1.PersistentVolume) error {
+	// Execute exportfs
+	cmd := exec.Command("exportfs", "-r")
+	out, err := cmd.CombinedOutput()
+	if err != nil {
+		return fmt.Errorf("exportfs -r failed with error: %v, output: %s", err, out)
+	}
+
+	return nil
+}
+
+type kernelExportBlockCreator struct{}
+
+var _ exportBlockCreator = &kernelExportBlockCreator{}
+
+// CreateBlock creates the text block to add to the /etc/exports file.
+func (e *kernelExportBlockCreator) CreateExportBlock(exportId, path string) string {
+	return "\n" + path + " *(rw,insecure,root_squash,fsid=" + exportId + ")\n"
+}
